% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trackhub.R
\name{easyTrackHub}
\alias{easyTrackHub}
\title{Building UCSC trackhubs}
\usage{
easyTrackHub(
  file_info,
  trackhub_name,
  trackhub_path,
  maintainer_email,
  is_multiwig = FALSE,
  copy_data = FALSE,
  ...
)
}
\arguments{
\item{file_info}{A data frame with the columns "sample_name", "file_path", "file_format", "reference_genome" and "data_type".}

\item{trackhub_name}{The identifier of the trackhub.}

\item{trackhub_path}{The path where to write the new trackhub.}

\item{maintainer_email}{A character string with an e-mail address.}

\item{is_multiwig}{Logical indicating if track should be created as multwig overlay.}

\item{copy_data}{Logical indicating whether to copy the data or to create symbolic links to the data.}

\item{...}{Additional parameters to be passed as UCSC parameters.}
}
\description{
This function inputs a data frame containing file paths to genomic coverage files
and their corresponding metadata. The function reorganizes the information and writes a directory
with the structure of a UCSC trackhub.
}
\examples{
library(easyTrackHubs)
data( test_df, package="easyTrackHubs" )
package_dir <- system.file("extdata", package="easyTrackHubs", mustWork=TRUE)
bigwig_files <- list.files( package_dir, full.names=TRUE )
names(bigwig_files) <- gsub(".bigWig", "", basename(bigwig_files))
test_df$file_path <- bigwig_files[test_df$sample_name]
easyTrackHub( test_df, trackhub_name="mth", trackhub_path="mth4", maintainer_email="prueba@novartis.com" )

}
